moved gtm logic into page component
