+++ client/public/index.html
@@ -62,7 +62,7 @@
 
       Note that we insert the GTM script tag from JS to ensure that
       GTM only initializes itself once we've set the title tag.
-      For more details, see GoogleTagManager.tsx in this repository.
+      For more details, see Page.tsx in this repository.
     -->
     <meta name="jf-gtm-id" content="UA-67069242-5">
     <script>
